backend/requirements.txt
absl-py==2.3.0
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==2.3.2
flatbuffers==25.2.10
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.72.1
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaggle==1.7.4.5
keras==2.14.0
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.1.0
narwhals==1.42.0
numpy==2.1.3
oauthlib==3.3.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
protobuf==4.25.8
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
scipy==1.15.3
simsimd==6.4.9
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
stringzilla==3.12.5
tenacity==9.1.2
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
text-unidecode==1.3
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.14.1
